Removed unnecessary Reference Model files
