+++ Melin.Server/Controllers/AuthController.cs
@@ -4,7 +4,6 @@ using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Serilog;
-using Task = Melin.Server.Models.Task;
 
 namespace Melin.Server.Controllers;
