+++ Melin.Server/Controllers/GroupController.cs
@@ -9,7 +9,6 @@ using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Serilog;
-using Task = Melin.Server.Models.Task;
 
 namespace Melin.Server.Controllers;
