--- Melin.Server/Melin.Server.http
@@ -1,6 +1,0 @@
-@Melin.Server_HostAddress = http://localhost:5018
-
-GET {{Melin.Server_HostAddress}}/api/
-Accept: application/json
-
-###
